+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - add `expire_date` and `data` fields to `Session` model.
+### Changed
+- Use `djplus.toml` file instead of `djplus.json` file to manage configurations.
+
 
 ## [2.0.0] - 2022-07-25
 ### Added