+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- add initial app called `blog`.
 - add `expire_date` and `data` fields to `Session` model.
 ### Changed
 - Use `djplus.toml` file instead of `djplus.json` file to manage configurations.