+++ CHANGELOG.md
@@ -5,12 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2022-10-27
 ### Added
 - add initial app called `blog`.
 - add `expire_date` and `data` fields to `Session` model.
 ### Changed
 - Use `djplus.toml` file instead of `djplus.json` file to manage configurations.
-
 
 ## [2.0.0] - 2022-07-25
 ### Added
@@ -43,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2022-04-09
 Initial release
 
+[2.1.0]: https://github.com/githashem/djplus/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/githashem/djplus/compare/v1.1.1...v2.0.0
 [1.1.1]: https://github.com/githashem/djplus/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/githashem/djplus/compare/v1.0.0...v1.1.0